Extract App Coordinator Tests
