+++ AnotherFeedlyAppTests/AnotherFeedlyAppTests.swift
@@ -1,28 +1,6 @@
 import XCTest
 @testable import AnotherFeedlyApp
 
-class AppCoordinatorTests: XCTestCase {
-
-    override func setUp() {
-        super.setUp()
-        // do setup
-    }
-
-    override func tearDown() {
-        //tear down
-        super.tearDown()
-    }
-
-    func testCoordinatorConfiguresRootVC() {
-        let window = UIWindow(frame: UIScreen.main.bounds)
-        let coordinator = AppCoordinator(window: window)
-
-        coordinator.appLaunched(options: nil)
-
-        XCTAssertNotNil(window.rootViewController)
-    }
-}
-
 class MockDelegate: NSObject, UIWebViewDelegate {
     var loadStarted: (() -> Void)
     init(loadStarted: @escaping (() -> Void)) {
